Add files via upload
This is a database for Restaurant Management System using SQL. Here we have taken some assumptions such as this restaurant accepts both online and offline payment. Cash on delivery is considered to be online payment and the bill paid in the restaurant is considered to be offline payment. Here the customer can make 1 order at a time and it can consist of only 3 items at a time. That's the limitation of this project. We can overcome this by taking 1 Item and then making it as multi-valued but then we have to normalize it further to remove the insert,delete and update anomalies by creating it into a new table !!